Move approvals to feature file
