+++ frame/nfts/src/lib.rs
@@ -1120,55 +1120,17 @@ pub mod pallet {
 			delegate: AccountIdLookupOf<T>,
 			maybe_deadline: Option<<T as SystemConfig>::BlockNumber>,
 		) -> DispatchResult {
-			ensure!(
-				Self::is_pallet_feature_enabled(PalletFeature::Approvals),
-				Error::<T, I>::MethodDisabled
-			);
-			let maybe_check: Option<T::AccountId> = T::ForceOrigin::try_origin(origin)
+			let maybe_check_origin = T::ForceOrigin::try_origin(origin)
 				.map(|_| None)
 				.or_else(|origin| ensure_signed(origin).map(Some).map_err(DispatchError::from))?;
-
 			let delegate = T::Lookup::lookup(delegate)?;
-
-			let mut details =
-				Item::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::UnknownItem)?;
-
-			let collection_config = Self::get_collection_config(&collection)?;
-			ensure!(
-				collection_config.is_setting_enabled(CollectionSetting::TransferableItems),
-				Error::<T, I>::ItemsNonTransferable
-			);
-
-			let collection_config = Self::get_collection_config(&collection)?;
-			ensure!(
-				collection_config.is_setting_enabled(CollectionSetting::TransferableItems),
-				Error::<T, I>::ItemsNonTransferable
-			);
-
-			if let Some(check) = maybe_check {
-				let is_admin = Self::has_role(&collection, &check, CollectionRole::Admin);
-				let permitted = is_admin || check == details.owner;
-				ensure!(permitted, Error::<T, I>::NoPermission);
-			}
-
-			let now = frame_system::Pallet::<T>::block_number();
-			let deadline = maybe_deadline.map(|d| d.saturating_add(now));
-
-			details
-				.approvals
-				.try_insert(delegate.clone(), deadline)
-				.map_err(|_| Error::<T, I>::ReachedApprovalLimit)?;
-			Item::<T, I>::insert(&collection, &item, &details);
-
-			Self::deposit_event(Event::ApprovedTransfer {
+			Self::do_approve_transfer(
+				maybe_check_origin,
 				collection,
 				item,
-				owner: details.owner,
 				delegate,
-				deadline,
-			});
-
-			Ok(())
+				maybe_deadline,
+			)
 		}
 
 		/// Cancel one of the transfer approvals for a specific item.
@@ -1193,43 +1155,11 @@ pub mod pallet {
 			item: T::ItemId,
 			delegate: AccountIdLookupOf<T>,
 		) -> DispatchResult {
-			let maybe_check: Option<T::AccountId> = T::ForceOrigin::try_origin(origin)
+			let maybe_check_origin = T::ForceOrigin::try_origin(origin)
 				.map(|_| None)
 				.or_else(|origin| ensure_signed(origin).map(Some).map_err(DispatchError::from))?;
-
 			let delegate = T::Lookup::lookup(delegate)?;
-
-			let mut details =
-				Item::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::UnknownItem)?;
-
-			let maybe_deadline =
-				details.approvals.get(&delegate).ok_or(Error::<T, I>::NotDelegate)?;
-
-			let is_past_deadline = if let Some(deadline) = maybe_deadline {
-				let now = frame_system::Pallet::<T>::block_number();
-				now > *deadline
-			} else {
-				false
-			};
-
-			if !is_past_deadline {
-				if let Some(check) = maybe_check {
-					let is_admin = Self::has_role(&collection, &check, CollectionRole::Admin);
-					let permitted = is_admin || check == details.owner;
-					ensure!(permitted, Error::<T, I>::NoPermission);
-				}
-			}
-
-			details.approvals.remove(&delegate);
-			Item::<T, I>::insert(&collection, &item, &details);
-			Self::deposit_event(Event::ApprovalCancelled {
-				collection,
-				item,
-				owner: details.owner,
-				delegate,
-			});
-
-			Ok(())
+			Self::do_cancel_approval(maybe_check_origin, collection, item, delegate)
 		}
 
 		/// Cancel all the approvals of a specific item.
@@ -1252,28 +1182,10 @@ pub mod pallet {
 			collection: T::CollectionId,
 			item: T::ItemId,
 		) -> DispatchResult {
-			let maybe_check: Option<T::AccountId> = T::ForceOrigin::try_origin(origin)
+			let maybe_check_origin = T::ForceOrigin::try_origin(origin)
 				.map(|_| None)
 				.or_else(|origin| ensure_signed(origin).map(Some).map_err(DispatchError::from))?;
-
-			let mut details =
-				Item::<T, I>::get(&collection, &item).ok_or(Error::<T, I>::UnknownCollection)?;
-
-			if let Some(check) = maybe_check {
-				let is_admin = Self::has_role(&collection, &check, CollectionRole::Admin);
-				let permitted = is_admin || check == details.owner;
-				ensure!(permitted, Error::<T, I>::NoPermission);
-			}
-
-			details.approvals.clear();
-			Item::<T, I>::insert(&collection, &item, &details);
-			Self::deposit_event(Event::AllApprovalsCancelled {
-				collection,
-				item,
-				owner: details.owner,
-			});
-
-			Ok(())
+			Self::do_clear_all_transfer_approvals(maybe_check_origin, collection, item)
 		}
 
 		/// Alter the attributes of a given collection.
